+++ packages/core/minos-microservice-aggregate/minos/aggregate/events/__init__.py
@@ -1,6 +1,14 @@
 from .entries import (
     EventEntry,
 )
+from .fields import (
+    FieldDiff,
+    FieldDiffContainer,
+    IncrementalFieldDiff,
+)
+from .models import (
+    Event,
+)
 from .repositories import (
     EventRepository,
     InMemoryEventRepository,