+++ packages/core/minos-microservice-saga/HISTORY.md
@@ -154,3 +154,18 @@ History
 * Be compatible with `minos-microservice-common~=0.4.0`.
 * Be compatible with `minos-microservice-aggregate~=0.4.0`.
 * Be compatible with `minos-microservice-networks~=0.4.0`.
+
+0.4.1 (2022-01-31)
+------------------
+
+* Update `README.md`.
+
+0.5.0 (2022-02-03)
+------------------
+
+* Minor changes.
+
+0.5.1 (2022-02-03)
+------------------
+
+* Fix bug related with dependency specification.